Change price to BigDecimal

+++ services/booking_service/src/main/java/com/demadev/model/SalonReport.java
@@ -2,6 +2,8 @@ package com.demadev.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class SalonReport {
 
@@ -10,5 +12,5 @@ public class SalonReport {
     private Double totalEarning;
     private Integer totalBookings;
     private Integer cancelledBookings;
-    private Double totalRefund;
+    private BigDecimal totalRefund;
 }
